feature resetet
