+++ src/health-management/pom.xml
@@ -137,12 +137,7 @@
 			<artifactId>spring-boot-starter-camunda-tasklist</artifactId>
 			<version>8.6.3</version>
 		</dependency>
-		<dependency>
-			<groupId>org.projectlombok</groupId>
-			<artifactId>lombok</artifactId>
-			<version>1.18.24</version>
-			<scope>provided</scope>
-		</dependency>
+
 	</dependencies>
