+++ src/health-management/src/main/java/com/example/health_management/HealthManagementApplication.java
@@ -1,14 +1,13 @@
-package com.example.health_management.app;
+package com.example.health_management;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
 
 import io.camunda.zeebe.spring.client.annotation.Deployment;
 
 @SpringBootApplication
 @Deployment(resources = "classpath*:/bpmn/**/")
-@ComponentScan(basePackages={"appointment", "healthrecord", "prescription", "task", "user"})
+
 public class HealthManagementApplication {
 
 	public static void main(String[] args) {
