+++ src/health-management/src/main/resources/db/migration/V1__init.sql
@@ -1,28 +1,20 @@
+
+
 CREATE TABLE health_records (
-                            id SERIAL PRIMARY KEY,
-                            username VARCHAR(10) NOT NULL UNIQUE,
-                            allergies VARCHAR(255),
-                            chronic_conditions VARCHAR(255),
-                            surgeries VARCHAR(255),
-                            health_insurance VARCHAR(255),
-                            FOREIGN KEY (username) REFERENCES public.user_entity(username) ON DELETE CASCADE
-                            );
+                                id SERIAL PRIMARY KEY,
+                                username VARCHAR(10) NOT NULL UNIQUE,
+                                allergies VARCHAR(255),
+                                chronic_conditions VARCHAR(255),
+                                surgeries VARCHAR(255),
+                                health_insurance VARCHAR(255),
+                                FOREIGN KEY (username) REFERENCES public.user_entity(username) ON DELETE CASCADE
+);
 
 CREATE TABLE visits (
-                    id SERIAL PRIMARY KEY,
-                    health_record_id BIGINT,
-                    date_of_visit DATE NOT NULL,
-                    diagnosis VARCHAR(255),
-                    treatment VARCHAR(255),
-                    FOREIGN KEY (health_record_id) REFERENCES health_records(id) ON DELETE CASCADE
-                    );
-
-CREATE TABLE appointment (
                         id SERIAL PRIMARY KEY,
-                        userID varchar(36),
-                        month varchar(255),
-                        day varchar(255),
-                        date DATE,
-                        time TIME,
-                        FOREIGN KEY (userID) REFERENCES public.user_entity(id) ON DELETE CASCADE
-                        );
+                        health_record_id BIGINT,
+                        date_of_visit DATE NOT NULL,
+                        diagnosis VARCHAR(255),
+                        treatment VARCHAR(255),
+                        FOREIGN KEY (health_record_id) REFERENCES health_records(id) ON DELETE CASCADE
+);
